modified courseLock to allow deselection
Moved the lockedCourse logic into redux instead of in the component
itself. There was a bug in the last commit that allowed you to change
the locked course without deselecting the current one, so if you tried
to lock the first course again, it would change its state to unlocked
because it was never deselected.