changed naming of selectedCourse to previewCourse
This is a better name for the concept it represents. Clicking on the
course titles at the top should give a preview of the course times that
are available. As mentioned in the new design.md document, the preview
should be an independent function of locking course sections. This
commit ensures that functionality, and the code that governs locking
coruses in was simplified at the same time.